Add support for sparse ext4 image creation.
This adds the -s option to the make_ext4fs tool, which now creates
"sparse" filesystem images, which is very useful for the large 32 Gbyte
filesystems we are now building.
This check-in also fixes make_ext4fs to properly create filesystems
larger thatn 4 Gbytes on 64-bit Linux, 32-bit android and Macs.

+++ ext4_utils/backed_block.c
@@ -113,9 +113,9 @@ void for_each_data_block(data_block_callback_t data_func,
 		last_block = db->block + DIV_ROUND_UP(db->len, info.block_size) - 1;
 
 		if (db->filename)
-			file_func(out, db->block * info.block_size, db->filename, db->offset, db->len);
+			file_func(out, (u64)db->block * info.block_size, db->filename, db->offset, db->len);
 		else
-			data_func(out, db->block * info.block_size, db->data, db->len);
+			data_func(out, (u64)db->block * info.block_size, db->data, db->len);
 	}
 }
